requirements.txt
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-keyvault-secrets==4.8.0
azure-mgmt-core==1.4.0
azure-mgmt-resource==23.1.1
azure-mgmt-sql==3.0.1
azure-mgmt-web==7.3.0
azure-mgmt-monitor==6.0.2
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==43.0.0
idna==3.7
isodate==0.6.1
msal==1.30.0
msal-extensions==1.2.0
msrest==0.7.1
oauthlib==3.2.2
portalocker==2.10.1
pycparser==2.22
PyJWT==2.9.0
pyodbc==5.1.0
requests==2.32.3
requests-oauthlib==2.0.0
six==1.16.0
typing_extensions==4.12.2
urllib3==2.2.2
